Don't abort connection after sending association release request
Don't set the completion flag in DicomClient after sending the association release request.
This causes the connection to be closed in some scenarios (e.g. when forceDisconnect is true in SendAsync)
which triggers socket exceptions on either the server side or the client side when the association release response is coming in.